Directory-Listing-Patch von eremit eingefuegt Ein paar kleinere Bugs bei unlimited Ressourcen fuer den Admin gefixed
git-svn-id: file:///var/svn/trunk@14 45fdb5c4-e40b-0410-b369-9aab4fe9a275

+++ syscp/customer_extras.php
@@ -48,21 +48,24 @@
 				$row['path']=str_replace($userinfo['documentroot'],'',$row['path']);
 				eval("\$htpasswds.=\"".getTemplate("extras/htpasswds_htpasswd")."\";");
 			}
-			eval("\$htpasswds.=\"".getTemplate("extras/htpasswds_addhtpasswd")."\";");
 			eval("echo \"".getTemplate("extras/htpasswds")."\";");
 		}
 
 		elseif($action=='delete' && $id!=0)
 		{
 			$result=$db->query_first("SELECT `id`, `customerid`, `username`, `path` FROM `".TABLE_PANEL_HTPASSWDS."` WHERE `customerid`='".$userinfo['customerid']."' AND `id`='$id'");
-				if(isset($_POST['send']) && $_POST['send']=='send')
+			if(isset($result['username']) && $result['username']!='')
 			{
-				$db->query("DELETE FROM `".TABLE_PANEL_HTPASSWDS."` WHERE `customerid`='".$userinfo['customerid']."' AND `id`='$id'");
-				inserttask('3',$result['path']);
-				header("Location: $filename?page=$page&s=$s");
-			}
-			else {
-				ask_yesno('extras_reallydelete', $filename, "id=$id;page=$page;action=$action");
+				if(isset($_POST['send']) && $_POST['send']=='send')
+				{
+					$db->query("DELETE FROM `".TABLE_PANEL_HTPASSWDS."` WHERE `customerid`='".$userinfo['customerid']."' AND `id`='$id'");
+					inserttask('3',$result['path']);
+					header("Location: $filename?page=$page&s=$s");
+				}
+				else
+				{
+					ask_yesno('extras_reallydelete', $filename, "id=$id;page=$page;action=$action");
+				}
 			}
 		}
 
@@ -136,4 +139,96 @@
 		}
 	}
 
+	elseif($page=='htaccess')
+	{
+		if($action=='')
+		{
+			$result=$db->query("SELECT * FROM `".TABLE_PANEL_HTACCESS."` WHERE `customerid`='".$userinfo['customerid']."'");
+			$htaccess = '';
+			while($row=$db->fetch_array($result))
+			{
+				$row['path']=str_replace($userinfo['documentroot'],'',$row['path']);
+				$row['options_indexes'] = str_replace('1', $lng['panel']['yes'], $row['options_indexes']);
+				$row['options_indexes'] = str_replace('0', $lng['panel']['no'], $row['options_indexes']);
+				eval("\$htaccess.=\"".getTemplate("extras/htaccess_htaccess")."\";");
+			}
+			eval("echo \"".getTemplate("extras/htaccess")."\";");
+		}
+		elseif($action=='delete' && $id!=0)
+		{
+			$result = $db->query_first("SELECT * FROM `".TABLE_PANEL_HTACCESS."` WHERE `customerid`='".$userinfo['customerid']."' AND `id`='$id'");
+			if(isset($result['customerid']) && $result['customerid']!='' && $result['customerid'] == $userinfo['customerid'])
+			{
+				if (isset($_POST['send']) && $_POST['send']=='send')
+				{
+					$db->query("DELETE FROM `".TABLE_PANEL_HTACCESS."` WHERE `customerid`='".$userinfo['customerid']."' AND `id`='$id'");
+					inserttask('3', $result['path']);
+					header("Location: $filename?page=$page&s=$s");
+				}
+				else
+				{
+					ask_yesno('extras_reallydelete_pathoptions', $filename, "id=$id;page=$page;action=$action");
+				}
+			}
+		}
+		elseif($action=='add')
+		{
+			if(isset($_POST['send']) && $_POST['send']=='send')
+			{
+				$path=makeCorrectDir(addslashes($_POST['path']));
+				$path=$userinfo['documentroot'].$path;
+				$path_dupe_check = $db->query_first("SELECT `id`, `path` FROM `".TABLE_PANEL_HTACCESS."` WHERE `path`='$path' AND `customerid`='".$userinfo['customerid']."'");
+				if ($path=='' || $path_dupe_check['path'] == $path || !is_dir($path))
+				{
+					if (!is_dir($path))
+					{
+						standard_error('directorymustexist');
+					}
+					else
+					{
+						standard_error('notallreqfieldsorerrors');
+					}
+				}
+				else
+				{
+					$db->query("INSERT INTO `".TABLE_PANEL_HTACCESS."` (`customerid`, `path`, `options_indexes`) VALUES ('".$userinfo['customerid']."', '$path', '".$_POST['options_indexes']."')");
+					inserttask('3',$path);
+					header("Location: $filename?page=$page&s=$s");
+				}
+			}
+			else 
+			{
+				$options_indexes = makeyesno('options_indexes','1','0','1');
+				eval("echo \"".getTemplate("extras/htaccess_add")."\";");
+			}
+		}
+		elseif($action=='edit' && $id!=0)
+		{
+			$result = $db->query_first("SELECT * FROM `".TABLE_PANEL_HTACCESS."` WHERE `customerid`='".$userinfo['customerid']."' AND `id`='$id'");
+			if(isset($result['customerid']) && $result['customerid']!='' && $result['customerid'] == $userinfo['customerid'])
+			{
+				if (isset($_POST['send']) && $_POST['send']=='send')
+				{
+					$option_indexes = intval($_POST['options_indexes']);
+					if($option_indexes != '1')
+					{
+						$option_indexes = '0';
+					}
+					if($option_indexes != $result['options_indexes'])
+					{
+						inserttask('3', $result['path']);
+					}
+					$db->query("UPDATE `".TABLE_PANEL_HTACCESS."` SET `options_indexes`='$option_indexes' WHERE `customerid`='".$userinfo['customerid']."' AND `id`='$id'");
+					inserttask('3',$result['path']);
+					header("Location: $filename?page=$page&s=$s");
+				}
+				else
+				{
+					$result['path'] = str_replace($userinfo['documentroot'], '', $result['path']);
+					$options_indexes = makeyesno('options_indexes', '1', '0', $result['options_indexes']);
+					eval("echo \"".getTemplate("extras/htaccess_edit")."\";");
+				}
+			}
+		}
+	}
 ?>
